Moved tests over to a different file
This is to reduce all the noise in the actual
collector file and make it easier to read.

Also hacks around an issue where roots are just
numbers in test but they are root? types in the
real world. I'll look around for a correct way
to do this but I'm not super sure at the moment.